requirements.txt
accelerate==1.12.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
asttokens==3.0.1
async-timeout==4.0.3
attrs==25.4.0
av==16.0.1
backoff==2.2.1
beautifulsoup4==4.14.3
black==25.11.0
blinker==1.9.0
brotli==1.2.0
bs4==0.0.2
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.2
croniter==6.0.0
cryptography==46.0.3
ctranslate2==4.6.2
cycler==0.12.1
dataclasses-json==0.6.7
ddgs==9.10.0
debugpy==1.8.17
decorator==5.2.1
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
dotenv==0.9.9
effdet==0.4.1
emoji==2.15.0
exceptiongroup==1.3.1
executing==2.2.1
fake-useragent==2.2.0
fastapi==0.121.3
faster-whisper==1.2.1
ffmpeg-python==0.2.0
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2024.12.0
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.2
google-cloud-vision==3.11.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
json_repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kagglehub==0.3.13
kiwisolver==1.4.9
langchain==1.2.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.1
langchain-experimental==0.4.1
langchain-huggingface==1.0.1
langchain-mongodb==0.8.0
langchain-neo4j==0.6.0
langchain-ollama==1.0.1
langchain-openai==1.1.0
langchain-qdrant==1.1.0
langchain-text-splitters==1.0.0
langdetect==1.0.9
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-checkpoint-mongodb==0.3.0
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.9
langsmith==0.4.59
lark==1.3.1
llvmlite==0.46.0
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.8
matplotlib-inline==0.2.1
mem0ai==1.0.1
ml_dtypes==0.5.4
more-itertools==10.8.0
moviepy==2.2.1
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.13.0
neo4j==5.28.2
neo4j-graphrag==1.10.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.2
nodejs-wheel-binaries==24.12.0
numba==0.63.1
numpy==2.2.6
oauthlib==3.3.1
olefile==0.47
ollama==0.6.1
omegaconf==2.3.0
onnx==1.20.0
onnxruntime==1.23.2
open_clip_torch==3.2.0
openai==2.9.0
openai-whisper==20250625
opencv-python==4.12.0.88
orjson==3.11.5
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20251107
pi_heif==1.1.1
pikepdf==10.0.2
pillow==11.3.0
platformdirs==4.5.0
playwright==1.57.0
portalocker==3.2.0
posthog==7.0.1
primp==0.15.0
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycocotools==2.0.11
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
pymongo==4.15.5
pyparsing==3.2.5
pypdf==6.4.0
pypdfium2==5.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-iso639==2025.11.16
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytokens==0.3.0
pytube==15.0.0
pytubefix==10.3.6
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
qdrant-client==1.16.0
rank-bm25==0.2.2
RapidFuzz==3.14.3
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.30.0
rq==2.6.1
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sounddevice==0.5.3
soupsieve==2.8
SpeechRecognition==3.14.4
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.50.0
streamlit==1.52.1
sympy==1.14.0
tenacity==9.1.2
tesseract==0.1.3
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==1.0.22
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
types-PyYAML==6.0.12.20250915
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.18.21
unstructured-client==0.42.6
unstructured.pytesseract==0.3.15
unstructured_inference==1.1.2
uritemplate==4.2.0
urllib3==2.6.2
uuid_utils==0.12.0
uv==0.9.11
uvicorn==0.38.0
watchdog==6.0.0
wcwidth==0.2.14
webencodings==0.5.1
websockets==16.0
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
